Cleanup - remove vault specific code & cleanup new graphQL implementation
