+++ Sources/CorePayments/Networking/GraphQL/GraphQLHTTPPostBody.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+/// The GraphQL query and variable details encoded to be sent in the POST body of a HTTP request
+struct GraphQLHTTPPostBody: Encodable {
+    
+    let query: String
+    let variables: Data
+}
