+++ Sources/CorePayments/Networking/GraphQL/GraphQLRequest.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+/// :nodoc:
+public struct GraphQLRequest {
+    
+    let queryNameForURL: String?
+    let query: String
+    let variables: Data // Dictionary
+}
